Uploaded dataconnection files
Added new files for connecting to the postgres data base.  Need to edit the database.ini file with your username, database name, and password to get it to connect.  the data_gen.py has a function for each query as part of the benchmark suite.  Will work on getting Data_gen.py fleshed out in stages.