Update to Queries and Part2 for full test suite
Updated the partial index queries to use ten attribute.  Using two didn't provide anything other than a sequential scan.  Updated the creation of tables to be worked on.  Original method was not copying over the key and index's from golden tables in mysql.  Updated all queries to output text files in JSON format for easy importing and extraction of data.